fix zhu is larger always better loop, improve structure
