+++ src/preprocessing/preprocessing_prompts/few_shot_health_learners/large_language_models_few_shot_healthlearners.py
@@ -33,7 +33,7 @@ def few_shot_paper_preprocessor(
     task = info_dict.get("task", "unknown_task")
     dataset = info_dict.get("dataset_name", "unknown_dataset")
     model_id = info_dict.get("model_name", "unknown_model")
-    num_shots = 0 #TODO: info_dict.get("num_shots", 0) rout from kwargs
+    num_shots = 0  # TODO: info_dict.get("num_shots", 0) rout from kwargs
     mode = info_dict.get("mode", "train")
 
     logger.info(
@@ -227,9 +227,10 @@ def wrap_for_few_shot_template(input_text: str, label_text: str) -> str:
     }
 
     prompt = (
-    "This is an example. Analyze the following text and determine the most likely diagnosis.\n"
-    "Text:\n"
-    f"{input_text}"
-    "Result:\n"
-    f"{result_json}")
+        "This is an example. Analyze the following text and determine the most likely diagnosis.\n"
+        "Text:\n"
+        f"{input_text}"
+        "Result:\n"
+        f"{result_json}"
+    )
     return prompt.strip()
